pip_requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
lxml==4.8.0
Pillow==9.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
